Benchmarks (#289)
* Add a test for executing modules with the StandaloneEnv

This commit adds an integration test for receiving a message
against the spacetimedb-core

* Add a benchmark for executing WASM modules

The modules benchmark is a part of the bench crate. This commit
also adds test-helpers crate which includes some helpers for running
modules tests

* Refactor modules handling in testing

This commit changes the dir structure for the modules that we use
for testing. Now all of the WASM modules that we use for testing are
in the crates/modules directory

On top of that the testing utilities and integration tests are now in the
testing crate. This makes it easier to write utilities for integration
tests.

* Use binary handler in benchmark tests

We use a binary message handler to call reducers when using websockets,
which is the primary way the clients will interact with SpacetimeDB.
Because of that it's better to test binary handlers instead of JSON
text handlers. It probably won't result in any performance measurement changes,
but if there is any regression in how we handle protobuf messages we will
be able to catch it easier

* Benchmarks cleanup

* Enable vs sqlite3 benchmarks

* Remove hyperfine step, it's preinstalled now

* Fix path to bench executable

---------

Co-authored-by: Tyler Cloutier <cloutiertyler@users.noreply.github.com>