Fix bug with more than one people subscribing for updates (#353)
When a new client subscribes for updates it wasn't actually added to a
list of subscribers. The problem was that we create a temporary
subscription instance that we drop if another one already exists. The
code was adding a new subscriber to the temporary subscription and not
the existing one