`ReducerEvent` enum in Rust client SDK
With this commit, row callbacks (`on_insert`, `on_delete`, `on_update`)
in the Rust client SDK take an additional argument,
`Option<&ReducerEvent>`, where `ReducerEvent` is an enum
generated by the CLI's codegen
with a variant for each reducer defined in the module.

Having the SDK pass around an autogenerated type in this way
(without adding a bunch of `<ReducerEvent>` generic parameters everywhere)
requires storing the `ReducerEvent` in an `Arc<dyn Any>`.
This has the added tangential benefit of avoiding cloning the `ReducerEvent`
for each row callback.