Fix schema equivalence check when updating a module (#39)
`IndexDef` contains a table id, which is not yet known (i.e. zero) when
constructing `TableDef` from the module describers. But it /is/ known
for the schema obtained from the database catalog, so schemas compare
inequal even if they're structurally equivalent.

Fixed by just updating the proposed schema for known tables before
comparing. Also update the smoke test to contain an indexed column.