Serialize Address as u128 in SATS
Address is meant to be an IPv6 address in the future, which is
presumably the reason this was implemented to serialize in network byte
order. However, deserializing a (Rust) array does not appear to work.

This should probably be fixed, but there also doesn't seem to be a
_strong_ reason against just serializing as a u128, which is what this
patch does.