Improve release build times
Thin LTO should be about as good as far LTO, but much faster, and debug info is something that takes enormous amount of time to handle throughout build pipeline but something we don't usually need in release builds (symbol information should be enough if it's backtraces that we're interested in).

This cuts `cargo build --release` for the CLI from 9m 40s to 5m 20s on my machine.