SDK: Append `Args` suffix to generated reducer args structs
This reduces the liklihood of name collisions,
e.g. when a user has both a table `Guess` and a reducer `guess`.
Prior to this commit, in that case, the SDK would generate
a `struct guess::Guess` for the table, and
a `struct guess_reducer::Guess` for the reducer args.
Now, the latter is `struct guess_reducer::GuessArgs`.
This will allow the CLI's codegen to re-export all generated structs
from its main `mod.rs` file.