Add `identity/:id/verify` route to verify identity/token pairs
This commit adds a new route to the `/identity` HTTP block,
`/identity/:identity/verify`,
which requires a SpacetimeDB auth header.

If the token in the authorization header is valid
and corresponds to the identity in the path,
it returns `NO_CONTENT`.

If the token in the authorization header is valid
but does not correspond to the identity in the path,
it returns `BAD_REQUEST`.

If the token in the authorization header is invalid or unsupplied,
it returns `UNAUTHORIZED`.