Add `/identity/public-key` route to get ECDSA public key
Corollary to https://github.com/clockworklabs/SpacetimeDB/pull/155.

This commit adds a new endpoint to the `/identity` block,
which returns the public key STDB uses to decode JWTs.
This will allow clients, incl. the CLI and SDKs,
to verify multiple identity/token pairs without spamming STDB with requests.

The public key can also be used as a "fingerprint", to identify the server.
This will allow clients to determine:
- When multiple URIs or addresses refer to the same STDB instance.
- When a STDB instance has rotated its keys, so past identities become invalid.