Fix ABI check for module with split data
Wasm data segments are often split into individual elements to avoid long stretches of zeroes in the middle, in particular by tools like wasm-opt.

Because the ABI check was only verifying the first data segment, it was complaining that module was malformed when ABI was simply in a different data segments. This commit fixes that check.

While at it, also converted panics into Rusty errors.