Fix authentication for the TypeScript SDK
TypeScript SDK returns a token in an onConnect callback. The current
behaviour may be confusing when used with a short lived token, cause it
will return the short lived token, not a long lived. Thus is someone
decides to always save the token received from the callback, it will
expire after 30s.

This commit changes the behaviour to generate a new long lived token if
a short lived token is used