Collect timing information about WASM execution in reducers
This commit augments our logging of reducer execution time
by logging the time spent specifically executing WASM code,
in addition to the overall elapsed time.

This is accomplished by adding two time-related fields to `WasmInstanceEnv`,
which we interact with in four contexts:
- before calling into a reducer.
- after returning from a reducer.
- after entering a host function called by a reducer.
- before returning from a host function called by a reducer.

The timings are probably not terribly precise,
as we necessarily capture a small amount of overhead during calls and returns,
but they're good enough for our purposes,
i.e. knowing whether slow reducers are slow
because they spend a lot of time in the datastore,
or because they spend a lot of time doing WASM computations.