Collect iterator into chunks early
Initially we used iterator because we wanted to lazily pull values from the database when Wasm requests the next item, so that e.g. early stop works as expected and doesn't read too much data.

Unfortunately, we discovered that this leads to crashes when someone tries to iterate over the database and update/delete items inside the iteration since iteration itself held a lock.

We added a workaround by collecting the iterator early into a Vec, and then returning a newly created iterator to the user.

This fixed the issue, but at this point lazy iterator mechanism turned from an optimisation into pure overhead - if we want to collect Vec, we might as well do this early in the pipeline.

Initial results show 25%-30% speedup to module iteration benchmarks.