Flatten subscription query into an actual set
Although semantically dubious, it is apparently legal to request a
subscription of the form:

    [
        "SELECT * FROM a",
        "SELECT * FROM b; SELECT * FROM a",
    ]

That is, sidestep the list type in the request by specifying multiple
SQL statements in a single string. We would previously keep a nested
list and not deduplicate equal queries.

With this change, the above is flattened to:

    [
        "SELECT * FROM a",
        "SELECT * FROM b",
    ]