Simplify and document WasmInstanceEnv's interface
This abstracts out all the internals of WasmInstanceEnv, and provides
a slightly higher level API. In particular, we expect to potentially
add more state associated to a single reducer invocation, and it will
be nice to interact with that in general rather than per type of
state.

This also documents WasmInstanceEnv and its members.