fix(430): Incremental evaluation for semijoins
Fixes #430.

For inserts, under set semantics:

The formula for incrementally evaluating a join before a transaction
has been applied is as follows

{A+ join B} U {A join B+} U {A+ join B+}

The formula for incrementally evaluating a join after a transaction
has been applied is as follows

{A+ join B} U {A join B+}

For deletes, under set semantics:

The formula for incrementally evaluating a join before a transaction
has been applied is as follows

{A- join B} U {A join B-}

The formula for incrementally evaluating a join after a transaction
has been applied is as follows

{A- join B} U {A join B-} U {A- join B-}

Note that SpacetimeDB is operating under the latter model.
Evaluation after the updates have already been applied to the underlying tables.
Therefore the 2nd and 4th formulas apply.