core: Fix commit log concurrency
Previously, constructing the commit payload to append to the message log
was done without holding the lock on the latter. This meant that commits
could be written to the log out-of-order.

Indeed, this could be observed on deployed databases by virtue of
verifying the hash chain (the parent hash is computed in
`generate_commit`).

To fix this, the lock is now acquired immediately and held until the
message is written (and potentially fsync'ed).