core: Add truncation to message log
Adds a method `reset_to(offset)` to `MessageLog`, which allows to
truncate the log to the given message offset.

Note that this currently requires to iterate over the last segment until
the message offset is found -- we may in the future maintain an offset
index (a la Kafka) to be able to seek closer to the desired message
offset.

`reset_to` subsumes the unused `reset_hard` method, and is itself not
yet used (this will be added in a later patch).