refactor(compiler): Swap the index and probe sides of an index join
Before this change we would swap the index and probe sides if

(1) the index side was replaced with a delta table,
(2) the probe side had an index defined on the join field, and
(3) the index side was the return side.

After this change we swap if

(1) the index side has been replaced with a delta table, and
(2) the probe side has an index defined on the join field.

We no longer place restrictions on the rows that are returned.