Note correctness of `state_view::Iter` and add test in `relational_db`.
Tyler had concerns about the behavior of `state_view::Iter`
w.r.t. the edge case of a committed-deleted-inserted row,
since the new datastore maintains the invariant that
the committed and inserted tables are disjoint,
where the previous datastore did not.

This commit adds comments describing that invariant to the iterator,
and adds a test in `relational_db` that the external semantics are preserved.