Add notes for future MVCC implementors re: disjoint committed and insert tables
Tyler convinced me via video call that, in MVCC, it is no longer valid
to treat the committed and insert tables as disjoint,
i.e. to elide inserts of already-present rows.

Doing so is still valid, and is more performant, in the current locking model,
so we retain that behavior here.

This commit adds comments to relevant code paths so that, when we move to MVCC,
we do not forget to remove the optimization.