Remove `row_pk` from client API; hash rows not row_ids on client
This commit removes the `row_pk` field from `message TableRowOperation` in the client API.
This is the beginning of addressing issue #831.

This has several implications for clients:
- Because we no longer have a stable content-addressed ID for each row,
  the client is responsible for hashing rows itself.
- This means that generated types must be `Hash + Eq`.
- This means that generated types must use `sats::F32` or `sats::F64`
  rather than `f32` or `f64`,
  as the former are `Eq + Hash` while the latter are not.
  (Aside: it's stupid that Rust floats are not `Eq + Ord + Hash`,
   since IEEE-754 defines a total equality and total ordering on floats.)

Exposing `sats::F32` to users is ugly, but I can't think of a better design.

This commit also makes some minor formatting changes to Rust codegen,
since it was emitting code that rustc warned about.

Still outstanding:

- [ ] Remove row_pk from the JSON API.
- [ ] Avoid computing row_pk in the subscription engine.
- [ ] Update other SDKs.
  - [ ] C#
  - [ ] TypeScript
  - [ ] Python