core: More reliable detection of incompatible schema updates (#734)
Compare `TableSchema` instead of `TableDef`, which should reject all
modifications to a table. Note that the values require conversion for
both the module and database schema to be comparable.

With this change, updating a module which only adds tables or modifies
reducers will be possible, but any change (including indexes,
constraints, etc) to an existing table will fail the update.