HostThreadPool: don't run Rayon threads in Tokio blocking threads
Prior to this commit, we ran our Rayon threads
within the Tokio blocking threadpool.
This was bad for several reasons, of which the most obvious was
that it ignored the Rayon `thread_name`.

Our Rayon threads still need access to the Tokio runtime
(which I'm not super jazzed about; see comment),
so this commit adds a custom `spawn_handler`
which behaves like the Rayon default spawn handler,
but also enters the Tokio runtime.

This commit also adds a comment about how we're over-allocating our available parallelism,
and in the future we should not do that.