Remove row_pk from the client API
Re: https://github.com/clockworklabs/SpacetimeDB/pull/840

This commit updates the TypeScript SDK to no longer use the `row_pk` field
in the client API,
as that field no longer exists.

As in the other SDKs, we replace our use of the `row_pk`
with the serialized representation of the row,
as this saves our needing to have objects/dicts/hash-maps keyed on domain types.

Unlike the other SDKs, we support either the binary (protobuf) or JSON APIs.
When using the binary API, we convert the BSATN row to a string,
and use that as the `rowPk`.
When using the JSON API, we `JSON.stringify` the row itself, and use that as the `rowPk`.
The latter is ugly and not performant,
but we don't care because the JSON API is slow anyways.

This commit also removes some uses of `any` from the deserialization code,
because I wanted the compiler to double-check my work.