Avoid `Header::find_pos_by_name` in `eval_incr`.
Prior to this commit, `to_mem_table_with_op_type` would call `Header::find_pos_by_name`
to locate the `__op_type` column, if it existed.
This was slow, as `Header::find_pos_by_name` is a linear scan in increasing order,
and the `__op_type` column is always either last or not present,
so `to_mem_table_with_op_type` would traverse every column in the `Header`.
This happened during every incremental evaluation for every query.

With this commit, we rely on the fact that `__op_type` is always the last column if present,
and check only the last column of the header.